merge

+++ src/main/java/org/jibe77/hermanas/data/repository/PictureRepository.java
@@ -3,7 +3,7 @@ package org.jibe77.hermanas.data.repository;
 import org.jibe77.hermanas.data.entity.Picture;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PictureRepository extends CrudRepository<Picture, Integer> {
+public interface PictureRepository extends CrudRepository<Picture, Long> {
 
     Picture findByPath(String path);
 }
